blockchain/requirements.txt
web3==6.11.3
py-solc-x==1.1.1
eth-account==0.10.0
loguru==0.7.2
python-dotenv==1.0.0
